zzMerged game

See merge request s08-webmobile1-sub2/S08P12A706!115

+++ back-end/liargame/src/main/java/com/sixsense/liargame/api/controller/SubjectController.java
@@ -23,7 +23,7 @@ public class SubjectController {
     }
 
     @GetMapping("/{subjectId}/words")
-    public ResponseEntity<?> getWord(@PathVariable Long subjectId, Long roomId) {
+    public ResponseEntity<?> getWord(@PathVariable Long subjectId, Integer roomId) {
         subjectService.selectRandomWord(subjectId, roomId);
         return ResponseEntity.ok().build();
     }
